+++ shared/actions/chat2/index.js
@@ -4,7 +4,6 @@ import * as ConfigGen from '../config-gen'
 import * as Constants from '../../constants/chat2'
 import * as GregorGen from '../gregor-gen'
 import * as I from 'immutable'
-import fs from 'fs'
 import * as FsGen from '../fs-gen'
 import * as NotificationsGen from '../notifications-gen'
 import * as RPCChatTypes from '../../constants/types/rpc-chat-gen'
@@ -1735,28 +1734,6 @@ function* attachmentPreviewSelect(action: Chat2Gen.AttachmentPreviewSelectPayloa
   }
 }
 
-// Handle an image pasted into a conversation
-function* attachmentPasted(action: Chat2Gen.AttachmentPastedPayload) {
-  const {conversationIDKey, data} = action.payload
-  const outboxID = Constants.generateOutboxID()
-  const path = yield Saga.call(RPCChatTypes.localGetUploadTempFileRpcPromise, {
-    outboxID,
-    filename: 'paste.png',
-  })
-  const paths = [
-    {
-      path,
-      outboxID,
-    },
-  ]
-  fs.writeFileSync(path, data)
-  yield Saga.put(
-    RouteTreeGen.createNavigateAppend({
-      path: [{props: {conversationIDKey, paths}, selected: 'attachmentGetTitles'}],
-    })
-  )
-}
-
 // Upload an attachment
 function* attachmentsUpload(action: Chat2Gen.AttachmentsUploadPayload) {
   const {conversationIDKey, paths, titles} = action.payload
@@ -2689,7 +2666,6 @@ function* chat2Saga(): Saga.SagaGenerator<any, any> {
   yield Saga.safeTakeEvery(Chat2Gen.attachmentPreviewSelect, attachmentPreviewSelect)
   yield Saga.safeTakeEvery(Chat2Gen.attachmentDownload, attachmentDownload)
   yield Saga.safeTakeEvery(Chat2Gen.attachmentsUpload, attachmentsUpload)
-  yield Saga.safeTakeEvery(Chat2Gen.attachmentPasted, attachmentPasted)
 
   yield Saga.safeTakeEveryPure(Chat2Gen.sendTyping, sendTyping)
   yield Saga.safeTakeEveryPure(Chat2Gen.resetChatWithoutThem, resetChatWithoutThem)
